Add Movie Trailer & Enhance Detail Screen UI/UX
Integrate movie trailers from YouTube using Android View Compose.
Upgrade UI with Material Theme.
Join query to fetch extra movie details.

+++ app/src/main/java/com/example/themovieapp/data/source/local/room/moviedetails/entity/ReviewEntity.kt
@@ -35,7 +35,7 @@ data class ReviewEntity(
         content = content,
         url = url,
         updatedAt = updatedAt,
-        rating = rating,
+        rating = rating.toDouble(),
         avatarPath = avatarPath,
         createdAt = createdAt,
     )
